Separate reddit.com preprocessor / postprocessor
We use the postprocessor module in the preprocessor.
This complicates things and blocks from refactoring
preprocessor/postprocessor code.

We move the preprocessor code to
`internal/pkg/preprocessor/sitespecific/reddit/`.

There is only one tiny `reddit.IsRedditURL` function that needs to be
duplicated between the preprocessor and postprocessor.
This is a minor downside to the benefits of better code organization.
